Refactor: Remove SignIn page, add StatsCard, PersonCard, and SearchBar components, and update dashboard types
