+++ app/(dashboard)/manager/team/page.tsx
@@ -1,32 +1,14 @@
 "use client";
 
 import React, { useState, useEffect } from 'react';
-import { Search, Phone, MessageCircle, Users, UserCheck } from 'lucide-react';
+import { Phone, Users, UserCheck } from 'lucide-react';
 import axios from 'axios';
 import Loader from '@/components/Loader';
-import Error from '@/components/ErrorBox'
-
-/**
- * Interface for individual person data displayed in the UI
- */
-interface PersonData {
-    id: string;
-    name: string;
-    role: string;
-    callCount: number;
-    messageCount?: number;
-    email?: string;
-}
-
-/**
- * Interface for stats card data
- */
-interface StatsCardData {
-    value: number;
-    label: string;
-    icon: React.ComponentType<any>;
-    isHighlighted?: boolean;
-}
+import Error from '@/components/ErrorBox';
+import PersonList from '@/components/manager/PersonList';
+import { PersonData, StatsCardData } from '@/types/dashboard';
+import StatsCard from '@/components/manager/StatCard';
+import SearchBar from '@/components/manager/SearchBar';
 
 /**
  * API Response interface for auditors endpoint
@@ -259,227 +241,6 @@ function filterPeople(query: string, dashboardData: DashboardData): {
         counsellors: filteredCounsellors
     };
 }
-
-// Component interfaces
-interface StatsCardProps {
-    data: StatsCardData;
-}
-
-interface PersonCardProps {
-    person: PersonData;
-    showMessages?: boolean;
-}
-
-interface PersonListProps {
-    title: string;
-    people: PersonData[];
-    totalCount: number;
-    showMessages?: boolean;
-    isLoading?: boolean;
-}
-
-interface SearchBarProps {
-    onSearch: (query: string) => void;
-    placeholder?: string;
-    isLoading?: boolean;
-}
-
-/**
- * StatsCard component - displays metric cards at the top
- */
-const StatsCard: React.FC<StatsCardProps> = ({ data }) => {
-    const { value, label, icon: Icon } = data;
-
-    return (
-        <div
-            className={`
-                rounded-xl p-6 bg-qc-dark/10
-            `}
-        >
-            <div className="flex items-start justify-between">
-                <div className="flex-1">
-                    <div className="text-2xl md:text-3xl font-bold mb-2 text-gray-800">
-                        {value}
-                    </div>
-                    <div className="text-sm md:text-base font-medium text-gray-600">
-                        {label}
-                    </div>
-                </div>
-                <div className="ml-3">
-                    <Icon className="h-6 w-6 text-qc-accent" />
-                </div>
-            </div>
-        </div>
-    );
-};
-
-/**
- * SearchBar component - handles search functionality with instant filtering
- */
-const SearchBar: React.FC<SearchBarProps> = ({
-    onSearch,
-    placeholder = "Search counsellor or auditor by name...",
-    isLoading = false
-}) => {
-    const [searchQuery, setSearchQuery] = useState('');
-
-    const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        const value = e.target.value;
-        setSearchQuery(value);
-        // Call onSearch on every keystroke for instant filtering
-        onSearch(value);
-    };
-
-    const clearSearch = () => {
-        setSearchQuery('');
-        onSearch('');
-    };
-
-    return (
-        <div className="relative mb-8">
-            <div className="absolute inset-y-0 left-0 pl-4 flex items-center pointer-events-none">
-                <Search className="h-5 w-5 text-gray-400" />
-            </div>
-            <input
-                type="text"
-                value={searchQuery}
-                onChange={handleInputChange}
-                placeholder={placeholder}
-                disabled={isLoading}
-                className={`
-                    w-full pl-12 pr-16 py-4 rounded-xl border-0 
-                 shadow-sm text-gray-900 bg-qc-dark/10 focus:outline-none focus:ring-1 focus:ring-qc-accent transition-all duration-200 text-base
-                    ${isLoading ? 'opacity-50 cursor-not-allowed' : ''}
-                `}
-            />
-            {searchQuery && (
-                <button
-                    onClick={clearSearch}
-                    className="absolute right-4 top-1/2 transform -translate-y-1/2 text-gray-400 hover:text-gray-600 transition-colors"
-                >
-                    <svg className="h-5 w-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-                    </svg>
-                </button>
-            )}
-        </div>
-    );
-};
-
-/**
- * PersonCard component - displays individual person information
- */
-const PersonCard: React.FC<PersonCardProps> = ({ person, showMessages = false }) => {
-    const { name, role, callCount, messageCount } = person;
-
-    return (
-        <div className="flex items-center justify-between p-4 rounded-xl mb-3 transition-all duration-200 hover:shadow-md bg-gray-50">
-            {/* Left side - Avatar and info */}
-            <div className="flex items-center space-x-3">
-                <div className="w-12 h-12 rounded-full bg-qc-accent flex items-center justify-center text-white font-semibold">
-                    {name.charAt(0)}
-                </div>
-                <div>
-                    <div className="font-medium text-base text-gray-800">
-                        {name}
-                    </div>
-                    <div className="text-sm text-gray-600">
-                        {role}
-                    </div>
-                </div>
-            </div>
-
-            {/* Right side - Call and message counts */}
-            <div className="flex items-center space-x-4">
-                <div className="flex items-center space-x-1">
-                    <Phone className="h-4 w-4 text-qc-dark" />
-                    <span className="text-sm font-medium text-gray-800">
-                        {callCount}
-                    </span>
-                </div>
-                {showMessages && messageCount !== undefined && (
-                    <div className="flex items-center space-x-1">
-                        <MessageCircle className="h-4 w-4 text-qc-dark" />
-                        <span className="text-sm font-medium text-gray-800">
-                            {messageCount}
-                        </span>
-                    </div>
-                )}
-            </div>
-        </div>
-    );
-};
-
-/**
- * PersonList component - displays a list of people with header
- */
-const PersonList: React.FC<PersonListProps> = ({
-    title,
-    people,
-    totalCount,
-    showMessages = false,
-    isLoading = false
-}) => {
-    return (
-        <div className="rounded-xl p-6 bg-qc-dark/10">
-            {/* Header */}
-            <div className="flex items-center justify-between mb-6">
-                <div className="flex items-center space-x-3">
-                    <h2 className="text-xl md:text-2xl font-bold text-gray-800">
-                        {title}
-                    </h2>
-                    <span className="bg-qc-accent text-white px-3 py-1 rounded-full text-sm font-medium">
-                        {totalCount}
-                    </span>
-                </div>
-                <button className="text-sm font-medium text-qc-dark hover:text-blue-700 transition-colors underline">
-                    View more
-                </button>
-            </div>
-
-            {/* Loading state */}
-            {isLoading && (
-                <div className="space-y-3">
-                    {[...Array(3)].map((_, index) => (
-                        <div key={index} className="animate-pulse">
-                            <div className="flex items-center space-x-3 p-4 rounded-xl bg-gray-50">
-                                <div className="w-12 h-12 bg-gray-300 rounded-full"></div>
-                                <div className="flex-1 space-y-2">
-                                    <div className="h-4 bg-gray-300 rounded w-3/4"></div>
-                                    <div className="h-3 bg-gray-300 rounded w-1/2"></div>
-                                </div>
-                                <div className="flex space-x-2">
-                                    <div className="h-6 bg-gray-300 rounded w-8"></div>
-                                    {showMessages && <div className="h-6 bg-gray-300 rounded w-8"></div>}
-                                </div>
-                            </div>
-                        </div>
-                    ))}
-                </div>
-            )}
-
-            {/* People list */}
-            {!isLoading && (
-                <div className="space-y-2">
-                    {people.length > 0 ? (
-                        people.map((person) => (
-                            <PersonCard
-                                key={person.id}
-                                person={person}
-                                showMessages={showMessages}
-                            />
-                        ))
-                    ) : (
-                        <div className="text-center py-8 text-gray-500">
-                            <p className="text-lg">No {title.toLowerCase()} found</p>
-                            <p className="text-sm opacity-70">Try adjusting your search criteria</p>
-                        </div>
-                    )}
-                </div>
-            )}
-        </div>
-    );
-};
 
 /**
  * Main Dashboard component with intelligent caching
