+++ components/auditor/AuditDetails.tsx
@@ -1,0 +1,308 @@
+"use client";
+
+import React, { useState, useEffect, useRef } from 'react';
+import { Check, Flag, Pause, Play, ChevronRight } from 'lucide-react';
+import { AuditItem } from '@/types/dashboard';
+
+/**
+ * Props for the audit details panel
+ */
+interface AuditDetailsProps {
+    audit: AuditItem | null;
+    onApprove: (auditId: string, comments?: string, isFlag?: boolean, flagReasons?: string) => void;
+    onApproveLoading: boolean;
+}
+
+/**
+ * Audit details panel component
+ */
+const AuditDetails: React.FC<AuditDetailsProps> = ({
+    audit,
+    onApprove,
+    onApproveLoading
+}) => {
+    const [flagReason, setFlagReason] = useState('');
+    const [showFlagInput, setShowFlagInput] = useState(false);
+    const [comments, setComments] = useState('');
+    const [isPlaying, setIsPlaying] = useState(false);
+    const [currentTime, setCurrentTime] = useState(0);
+    const [audioDuration, setAudioDuration] = useState(0);
+    const [audioLoading, setAudioLoading] = useState(false);
+    const [audioError, setAudioError] = useState('');
+
+    const audioRef = useRef<HTMLAudioElement | null>(null);
+
+    useEffect(() => {
+        setComments('');
+        setFlagReason('');
+        setShowFlagInput(false);
+        setIsPlaying(false);
+        setCurrentTime(0);
+        setAudioDuration(0);
+        setAudioError('');
+
+        // Reset audio if audit changes
+        if (audioRef.current) {
+            audioRef.current.pause();
+            audioRef.current.currentTime = 0;
+        }
+    }, [audit]);
+
+    // Initialize audio when audit changes
+    useEffect(() => {
+        if (audit?.recordingUrl) {
+            const audio = new Audio();
+            audioRef.current = audio;
+
+            audio.addEventListener('loadstart', () => {
+                setAudioLoading(true);
+                setAudioError('');
+            });
+
+            audio.addEventListener('loadedmetadata', () => {
+                setAudioDuration(audio.duration);
+                setAudioLoading(false);
+            });
+
+            audio.addEventListener('timeupdate', () => {
+                setCurrentTime(audio.currentTime);
+            });
+
+            audio.addEventListener('ended', () => {
+                setIsPlaying(false);
+                setCurrentTime(0);
+            });
+
+            audio.addEventListener('error', () => {
+                setAudioLoading(false);
+                setAudioError('Failed to load audio');
+                setIsPlaying(false);
+            });
+
+            audio.src = audit.recordingUrl;
+            audio.preload = 'metadata';
+
+            return () => {
+                audio.pause();
+                audio.remove();
+            };
+        }
+    }, [audit?.recordingUrl]);
+
+    const togglePlayPause = () => {
+        if (!audioRef.current) return;
+
+        if (isPlaying) {
+            audioRef.current.pause();
+            setIsPlaying(false);
+        } else {
+            audioRef.current.play()
+                .then(() => {
+                    setIsPlaying(true);
+                })
+                .catch((error) => {
+                    setAudioError('Failed to play audio');
+                    setIsPlaying(false);
+                });
+        }
+    };
+
+    const handleProgressClick = (e: React.MouseEvent<HTMLDivElement>) => {
+        if (!audioRef.current || !audioDuration) return;
+
+        const rect = e.currentTarget.getBoundingClientRect();
+        const clickX = e.clientX - rect.left;
+        const width = rect.width;
+        const newTime = (clickX / width) * audioDuration;
+
+        audioRef.current.currentTime = newTime;
+        setCurrentTime(newTime);
+    };
+
+    const formatTime = (seconds: number) => {
+        if (isNaN(seconds)) return '0:00';
+        const minutes = Math.floor(seconds / 60);
+        const remainingSeconds = Math.floor(seconds % 60);
+        return `${minutes}:${remainingSeconds.toString().padStart(2, '0')}`;
+    };
+
+    const handleApproveClick = () => {
+        if (audit) {
+            onApprove(audit.id, comments.trim() || undefined, showFlagInput, flagReason.trim());
+        }
+    };
+
+    if (!audit) {
+        return (
+            <div className="bg-white rounded-lg border border-gray-200 p-6 h-full flex items-center justify-center">
+                <div className="text-center text-gray-500">
+                    <div className="w-16 h-16 bg-gray-100 rounded-full flex items-center justify-center mx-auto mb-4">
+                        <ChevronRight className="w-8 h-8 text-gray-400" />
+                    </div>
+                    <p>Select an audit to view details</p>
+                </div>
+            </div>
+        );
+    }
+
+    return (
+        <div className="bg-white rounded-lg border border-gray-200 p-4 sm:p-6 h-full flex flex-col">
+            <div className="flex-1 overflow-y-auto">
+                {/* Header */}
+                <div className="border-b border-gray-200 pb-4 mb-4">
+                    <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-2">
+                        <h2 className="text-lg font-semibold text-qc-primary">
+                            Current Review
+                        </h2>
+                        <div className="flex items-center gap-2">
+                            <span className="text-sm text-gray-600">Call ID:</span>
+                            <span className="font-medium text-qc-primary">{audit.callId}</span>
+                        </div>
+                    </div>
+                    <div className="flex items-center gap-2 mt-2">
+                        <span className="text-sm text-gray-600">AI Confidence:</span>
+                        <span className="font-medium text-qc-accent">{audit.confidence}%</span>
+                    </div>
+                </div>
+
+                {/* Call Recording */}
+                <div className="mb-6">
+                    <h3 className="font-semibold text-qc-primary mb-3">Call Recording</h3>
+                    <div className="p-3 bg-gray-50 rounded-lg">
+                        {audioError ? (
+                            <div className="text-red-600 text-sm mb-2">{audioError}</div>
+                        ) : null}
+
+                        <div className="flex items-center gap-3 mb-2">
+                            <button
+                                onClick={togglePlayPause}
+                                disabled={audioLoading || !!audioError || !audit.recordingUrl}
+                                className="w-8 h-8 bg-qc-accent text-white rounded-full flex items-center justify-center hover:bg-qc-dark transition-colors disabled:bg-gray-300 disabled:cursor-not-allowed"
+                            >
+                                {audioLoading ? (
+                                    <div className="w-4 h-4 border-2 border-white border-t-transparent rounded-full animate-spin"></div>
+                                ) : isPlaying ? (
+                                    <Pause className="w-4 h-4" />
+                                ) : (
+                                    <Play className="w-4 h-4" />
+                                )}
+                            </button>
+
+                            <div className="flex-1">
+                                <div
+                                    className="bg-gray-300 rounded-full h-2 cursor-pointer"
+                                    onClick={handleProgressClick}
+                                >
+                                    <div
+                                        className="bg-qc-accent h-2 rounded-full transition-all duration-100"
+                                        style={{
+                                            width: audioDuration > 0 ? `${(currentTime / audioDuration) * 100}%` : '0%'
+                                        }}
+                                    ></div>
+                                </div>
+                            </div>
+
+                            <span className="text-sm text-gray-600 min-w-[80px] text-right">
+                                {formatTime(currentTime)} / {formatTime(audioDuration || audit.duration * 60)}
+                            </span>
+                        </div>
+
+                        {!audit.recordingUrl && (
+                            <div className="text-gray-500 text-sm">No recording URL available</div>
+                        )}
+                    </div>
+                </div>
+
+                {/* Summary */}
+                <div className="mb-6">
+                    <h3 className="font-semibold text-qc-primary mb-3">Summary</h3>
+                    <div className="bg-gray-50 p-4 rounded-lg">
+                        <p className="text-sm text-gray-700 leading-relaxed">
+                            {audit.summary}
+                        </p>
+                    </div>
+                </div>
+
+                {/* Sentiments */}
+                <div className="mb-6">
+                    <h3 className="font-semibold text-qc-primary mb-3">Sentiments</h3>
+                    <span className={`inline-flex px-3 py-1 rounded-full text-sm font-medium ${audit.sentiments === 'Positive'
+                        ? 'bg-green-100 text-green-800'
+                        : audit.sentiments === 'Negative'
+                            ? 'bg-red-100 text-red-800'
+                            : 'bg-yellow-100 text-yellow-800'
+                        }`}>
+                        {audit.sentiments}
+                    </span>
+                </div>
+
+                {/* Anomalies */}
+                <div className="mb-6">
+                    <h3 className="font-semibold text-qc-primary mb-3">Anomalies</h3>
+                    <div className="bg-gray-50 p-4 rounded-lg">
+                        <p className="text-sm text-gray-700">
+                            {audit.anomalies}
+                        </p>
+                    </div>
+                </div>
+
+                {/* Comments */}
+                <div className="mb-6">
+                    <h3 className="font-semibold text-qc-primary mb-3">Comments</h3>
+                    <textarea
+                        placeholder="Write comment on audit..."
+                        className="w-full p-3 border border-gray-300 rounded-lg resize-none focus:ring-2 focus:ring-qc-accent focus:border-transparent"
+                        rows={3}
+                        value={comments}
+                        onChange={(e) => setComments(e.target.value)}
+                    />
+                </div>
+
+                {/* Flag Section */}
+                <div className="mb-6">
+                    <div className="flex items-center gap-2 mb-3">
+                        <button
+                            onClick={() => setShowFlagInput(!showFlagInput)}
+                            className="flex items-center gap-2 px-3 py-2 bg-red-50 text-red-700 rounded-lg hover:bg-red-100 transition-colors"
+                        >
+                            <Flag className="w-4 h-4" />
+                            Flag
+                        </button>
+                    </div>
+                    {showFlagInput && (
+                        <div className="space-y-3">
+                            <textarea
+                                placeholder="Write reason for flag..."
+                                value={flagReason}
+                                onChange={(e) => setFlagReason(e.target.value)}
+                                className="w-full p-3 border border-gray-300 rounded-lg resize-none focus:ring-2 focus:ring-qc-accent focus:border-transparent"
+                                rows={3}
+                            />
+                        </div>
+                    )}
+                </div>
+            </div>
+
+            {/* Action Buttons */}
+            <div className="flex pt-4 border-t border-gray-200">
+                <button
+                    onClick={handleApproveClick}
+                    disabled={onApproveLoading}
+                    className={`
+        flex-1 flex items-center justify-center gap-2 px-4 py-3 rounded-lg transition-colors
+        ${onApproveLoading ? 'bg-gray-400 cursor-not-allowed' : 'bg-qc-accent hover:bg-qc-dark text-white'}
+      `}
+                >
+                    <Check className="w-4 h-4" />
+                    {onApproveLoading ? (
+                        <span className="animate-pulse">Approving...</span>
+                    ) : (
+                        <span>Approve</span>
+                    )}
+                </button>
+            </div>
+        </div>
+    );
+};
+
+export default AuditDetails;
